`LinuxInhibitorTask`: Move code to a parent class
